+++ src/components/Avatar/Avatar.style.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { StyledAvatarProps } from "@/types/avatar";
+
+export const StyledImg = styled.img<StyledAvatarProps>`
+  ${(p) => p.sizeStyle};
+  width: var(--avatar-width, 20px);
+  height: var(--avatar-height, 20px);
+  border-radius: 99%;
+`;
