+++ src/components/Avatar/Avatar.variant.ts
@@ -1,0 +1,12 @@
+import { css } from "styled-components";
+
+export const SIZES = {
+  md: css`
+    --avatar-width: 20px;
+    --avatar-height: 20px;
+  `,
+  lg: css`
+    --avatar-width: 50px;
+    --avatar-height: 50px;
+  `,
+};
