+++ src/types/avatar.ts
@@ -1,0 +1,10 @@
+import { CSSProp } from "styled-components";
+
+export interface AvatarProps {
+  src: string;
+  size: 'md' | 'lg';
+}
+
+export interface StyledAvatarProps {
+  sizeStyle: CSSProp;
+}
